+++ docs/source/register/images/register_template.png
