+++ docs/source/register/index.rst
@@ -1,0 +1,7 @@
+Register
+========
+
+.. toctree::
+    :maxdepth: 1
+
+    ./endpoints
