+++ piccolo_api/register/endpoints.py
@@ -1,0 +1,191 @@
+from __future__ import annotations
+
+import os
+import re
+import typing as t
+from abc import ABCMeta, abstractproperty
+from json import JSONDecodeError
+
+from jinja2 import Environment, FileSystemLoader
+from piccolo.apps.user.tables import BaseUser
+from starlette.datastructures import URL
+from starlette.endpoints import HTTPEndpoint, Request
+from starlette.exceptions import HTTPException
+from starlette.responses import HTMLResponse, RedirectResponse
+from starlette.status import HTTP_303_SEE_OTHER
+
+if t.TYPE_CHECKING:  # pragma: no cover
+    from jinja2 import Template
+    from starlette.responses import Response
+
+
+SIGNUP_TEMPLATE_PATH = os.path.join(
+    os.path.dirname(os.path.dirname(__file__)), "templates", "register.html"
+)
+
+
+EMAIL_REGEX = re.compile(r"(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)")
+
+
+class RegisterEndpoint(HTTPEndpoint, metaclass=ABCMeta):
+    @abstractproperty
+    def _auth_table(self) -> t.Type[BaseUser]:
+        raise NotImplementedError
+
+    @abstractproperty
+    def _redirect_to(self) -> t.Union[str, URL]:
+        """
+        Where to redirect to after login is successful.
+        """
+        raise NotImplementedError
+
+    @abstractproperty
+    def _register_template(self) -> Template:
+        raise NotImplementedError
+
+    def render_template(
+        self, request: Request, template_context: t.Dict[str, t.Any] = {}
+    ) -> HTMLResponse:
+        # If CSRF middleware is present, we have to include a form field with
+        # the CSRF token. It only works if CSRFMiddleware has
+        # allow_form_param=True, otherwise it only looks for the token in the
+        # header.
+        csrftoken = request.scope.get("csrftoken")
+        csrf_cookie_name = request.scope.get("csrf_cookie_name")
+
+        return HTMLResponse(
+            self._register_template.render(
+                csrftoken=csrftoken,
+                csrf_cookie_name=csrf_cookie_name,
+                request=request,
+                **template_context,
+            )
+        )
+
+    async def get(self, request: Request) -> HTMLResponse:
+        return self.render_template(request)
+
+    async def post(self, request: Request) -> Response:
+        # Some middleware (for example CSRF) has already awaited the request
+        # body, and adds it to the request.
+        body = request.scope.get("form")
+
+        if not body:
+            try:
+                body = await request.json()
+            except JSONDecodeError:
+                body = await request.form()
+
+        username = body.get("username", None)
+        email = body.get("email", None)
+        password = body.get("password", None)
+        confirm_password = body.get("confirm_password", None)
+
+        if (
+            (not username)
+            or (not email)
+            or (not password)
+            or (not confirm_password)
+        ):
+            raise HTTPException(
+                status_code=401,
+                detail="Form is invalid. Missing one or more fields.",
+            )
+
+        if not EMAIL_REGEX.fullmatch(email):
+            if body.get("format") == "html":
+                return self.render_template(
+                    request,
+                    template_context={"error": "Invalid email address."},
+                )
+            else:
+                raise HTTPException(
+                    status_code=401, detail="Invalid email address."
+                )
+
+        if len(password) < 6:
+            if body.get("format") == "html":
+                return self.render_template(
+                    request,
+                    template_context={
+                        "error": "Password must be at least 6 characters long."
+                    },
+                )
+            else:
+                raise HTTPException(
+                    status_code=401,
+                    detail="Password must be at least 6 characters long.",
+                )
+
+        if confirm_password != password:
+            if body.get("format") == "html":
+                return self.render_template(
+                    request,
+                    template_context={"error": "Passwords do not match."},
+                )
+            else:
+                raise HTTPException(
+                    status_code=401, detail="Passwords do not match."
+                )
+
+        if await self._auth_table.count().where(
+            self._auth_table.email == email,
+            self._auth_table.username == username,
+        ):
+            if body.get("format") == "html":
+                return self.render_template(
+                    request,
+                    template_context={
+                        "error": "User with email or username already exists."
+                    },
+                )
+            else:
+                raise HTTPException(
+                    status_code=401,
+                    detail="User with email or username already exists.",
+                )
+
+        await self._auth_table.create_user(
+            username=username, password=password, email=email
+        )
+
+        return RedirectResponse(
+            url=self._redirect_to, status_code=HTTP_303_SEE_OTHER
+        )
+
+
+def register(
+    auth_table: t.Optional[t.Type[BaseUser]] = None,
+    redirect_to: t.Union[str, URL] = "/login/",
+    template_path: t.Optional[str] = None,
+) -> t.Type[RegisterEndpoint]:
+    """
+    An endpoint for register user.
+
+    :param auth_table:
+        Which table to create the user in. If not specified, it defaults to
+        :class:`BaseUser <piccolo.apps.user.tables.BaseUser>`.
+    :param redirect_to:
+        Where to redirect to after successful registration.
+    :param template_path:
+        If you want to override the default register HTML template, you can do
+        so by specifying the absolute path to a custom template. For example
+        ``'/some_directory/register.html'``. Refer to the default template at
+        ``piccolo_api/session_auth/templates/register.html`` as a basis for
+        your custom template.
+
+    """
+    template_path = (
+        SIGNUP_TEMPLATE_PATH if template_path is None else template_path
+    )
+
+    directory, filename = os.path.split(template_path)
+    environment = Environment(loader=FileSystemLoader(directory))
+    register_template = environment.get_template(filename)
+
+    class _RegisterEndpoint(RegisterEndpoint):
+        _auth_table = auth_table or BaseUser
+        _redirect_to = redirect_to
+        _register_template = register_template
+
+    return _RegisterEndpoint
