+++ piccolo_api/templates/session_logout.html
